ai/requirements.txt
streamlit
#google-generativeai
requests
pillow
ocrmac
